requirements.txt
python-telegram-bot==20.8
httpx>=0.24.0,<0.27.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
supabase==2.3.3 